plaqueGen/requirements.txt
pillow==7.1.2
# pycld2
pandas